squashed down components in screens
