--- frontend-dart/lib/core/app_core.dart
@@ -1,322 +1,0 @@
-// lib/core/app_core.dart
-import 'package:flutter/material.dart';
-
-class AppTheme {
-  static const primary = Color(0xFF1DB954);
-  static const background = Color(0xFF121212);
-  static const surface = Color(0xFF282828);
-  static const surfaceVariant = Color(0xFF333333);
-  static const onSurface = Color(0xFFFFFFFF);
-  static const onSurfaceVariant = Color(0xFFB3B3B3);
-  static const textSecondary = Color(0xFFB3B3B3);
-  static const error = Color(0xFFE91429);
-  static const success = Color(0xFF00C851);
-
-  static ThemeData get darkTheme => ThemeData(
-    useMaterial3: true,
-    brightness: Brightness.dark,
-    primaryColor: primary,
-    scaffoldBackgroundColor: background,
-    cardColor: surface,
-    colorScheme: const ColorScheme.dark(
-      primary: primary,
-      surface: surface,
-      background: background,
-      error: error,
-      onPrimary: Colors.white,
-      onSurface: onSurface,
-      onBackground: Colors.white,
-    ),
-    appBarTheme: const AppBarTheme(
-      backgroundColor: background,
-      foregroundColor: Colors.white,
-      elevation: 0,
-    ),
-    elevatedButtonTheme: ElevatedButtonThemeData(
-      style: ElevatedButton.styleFrom(
-        backgroundColor: primary,
-        foregroundColor: Colors.black,
-        minimumSize: const Size(88, 50),
-        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(25)),
-      ),
-    ),
-  );
-
-  static Widget buildCard({required Widget child, EdgeInsets? padding, EdgeInsets? margin, Color? color}) {
-    return Container(
-      padding: padding ?? const EdgeInsets.all(16),
-      margin: margin ?? const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
-      decoration: BoxDecoration(
-        color: color ?? surface,
-        borderRadius: BorderRadius.circular(12),
-      ),
-      child: child,
-    );
-  }
-
-  static Widget buildHeaderCard({required Widget child}) {
-    return Container(
-      padding: const EdgeInsets.all(24),
-      margin: const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
-      decoration: BoxDecoration(
-        gradient: LinearGradient(
-          begin: Alignment.topLeft,
-          end: Alignment.bottomRight,
-          colors: [primary, primary.withOpacity(0.8)],
-        ),
-        borderRadius: BorderRadius.circular(16),
-        boxShadow: [
-          BoxShadow(
-            color: primary.withOpacity(0.3),
-            blurRadius: 10,
-            offset: const Offset(0, 4),
-          ),
-        ],
-      ),
-      child: child,
-    );
-  }
-
-  static Widget buildFormCard({required Widget child, String? title, IconData? titleIcon}) {
-    return Card(
-      margin: const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
-      color: surface,
-      child: Padding(
-        padding: const EdgeInsets.all(16),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            if (title != null) ...[
-              Row(
-                children: [
-                  if (titleIcon != null) ...[
-                    Icon(titleIcon, color: primary, size: 20),
-                    const SizedBox(width: 8),
-                  ],
-                  Text(title, style: const TextStyle(fontSize: 18, fontWeight: FontWeight.bold, color: Colors.white)),
-                ],
-              ),
-              const SizedBox(height: 16),
-            ],
-            child,
-          ],
-        ),
-      ),
-    );
-  }
-
-  static Widget buildStandardCard({required Widget child}) {
-    return Card(
-      margin: const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
-      color: surface,
-      child: Padding(
-        padding: const EdgeInsets.all(16),
-        child: child,
-      ),
-    );
-  }
-
-  static Widget buildPrimaryButton({
-    required String text,
-    required VoidCallback? onPressed,
-    IconData? icon,
-    bool isLoading = false,
-  }) {
-    return SizedBox(
-      width: double.infinity,
-      height: 50,
-      child: ElevatedButton.icon(
-        onPressed: isLoading ? null : onPressed,
-        icon: isLoading 
-          ? const SizedBox(width: 16, height: 16, child: CircularProgressIndicator(strokeWidth: 2, color: Colors.black))
-          : Icon(icon ?? Icons.check, size: 16),
-        label: Text(text),
-        style: ElevatedButton.styleFrom(
-          backgroundColor: primary,
-          foregroundColor: Colors.black,
-          shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(25)),
-        ),
-      ),
-    );
-  }
-
-  static Widget buildSecondaryButton({
-    required String text,
-    required VoidCallback? onPressed,
-    IconData? icon,
-    bool fullWidth = true,
-  }) {
-    return SizedBox(
-      width: fullWidth ? double.infinity : null,
-      height: 50,
-      child: OutlinedButton.icon(
-        onPressed: onPressed,
-        icon: Icon(icon ?? Icons.check, size: 16),
-        label: Text(text),
-        style: OutlinedButton.styleFrom(
-          foregroundColor: primary,
-          side: const BorderSide(color: primary),
-          shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(25)),
-        ),
-      ),
-    );
-  }
-
-  static InputDecoration getInputDecoration({required String labelText, String? hintText, IconData? prefixIcon}) {
-    return InputDecoration(
-      labelText: labelText,
-      hintText: hintText,
-      prefixIcon: prefixIcon != null ? Icon(prefixIcon, color: onSurfaceVariant) : null,
-      filled: true,
-      fillColor: surfaceVariant,
-      border: OutlineInputBorder(borderRadius: BorderRadius.circular(8), borderSide: BorderSide.none),
-      focusedBorder: OutlineInputBorder(borderRadius: BorderRadius.circular(8), borderSide: const BorderSide(color: primary, width: 2)),
-    );
-  }
-}
-
-class AppSizes {
-  static const EdgeInsets screenPadding = EdgeInsets.all(16.0);
-  static const EdgeInsets cardPadding = EdgeInsets.all(16.0);
-  static const double borderRadius = 12.0;
-  static const double iconSize = 24.0;
-  static const double avatarRadius = 20.0;
-}
-
-class AppDurations {
-  static const Duration shortDelay = Duration(milliseconds: 100);
-  static const Duration mediumDelay = Duration(milliseconds: 300);
-  static const Duration longDelay = Duration(milliseconds: 500);
-  static const Duration animationDuration = Duration(milliseconds: 250);
-}
-
-class AppAnimations {
-  static const Curve defaultCurve = Curves.easeInOut;
-  static final Tween<double> fadeIn = Tween<double>(begin: 0.0, end: 1.0);
-  static final Tween<Offset> slideInFromBottom = Tween<Offset>(
-    begin: const Offset(0.0, 1.0),
-    end: Offset.zero,
-  );
-}
-
-class AppConstants {
-  static const String appName = 'Music Room';
-  static const String version = '1.0.0';
-  static const String defaultApiBaseUrl = 'http://localhost:8000';
-  static const String contentTypeJson = 'application/json';
-  static const String authorizationPrefix = 'Token';
-  static const int minPasswordLength = 8;
-}
-
-class AppRoutes {
-  static const String home = '/home';
-  static const String auth = '/auth';
-  static const String profile = '/profile';
-  static const String playlistEditor = '/playlist_editor';
-  static const String trackSearch = '/track_search';
-  static const String publicPlaylists = '/public_playlists';
-  static const String playlistSharing = '/playlist_sharing';
-  static const String trackSelection = '/track_selection';
-  static const String player = '/player';
-  static const String friends = '/friends';
-  static const String addFriend = '/add_friend';
-  static const String friendRequests = '/friend_requests';
-  static const String deviceManagement = '/device_management';
-  static const String controlDelegation = '/control_delegation';
-  static const String musicFeatures = '/music_features';
-  static const String trackVote = '/track_vote';
-  static const String deezerTrackDetail = '/deezer_track_detail';
-  static const String userPasswordChange = '/user_password_change';
-  static const String socialNetworkLink = '/social_network_link';
-}
-
-class AppStrings {
-  static const String signIn = 'Sign In';
-  static const String signUp = 'Sign Up';
-  static const String username = 'Username';
-  static const String email = 'Email';
-  static const String password = 'Password';
-  static const String login = 'Login';
-  static const String logout = 'Logout';
-  static const String loginSuccessful = 'Login successful';
-  static const String accountCreated = 'Account created successfully';
-  static const String confirmLogout = 'Are you sure you want to logout?';
-  
-  static const String save = 'Save';
-  static const String cancel = 'Cancel';
-  static const String confirm = 'Confirm';
-  static const String delete = 'Delete';
-  static const String add = 'Add';
-  static const String search = 'Search';
-  static const String ok = 'OK';
-  
-  static const String playlists = 'Playlists';
-  static const String publicPlaylists = 'Public Playlists';
-  static const String createPlaylist = 'Create Playlist';
-  static const String editPlaylist = 'Edit Playlist';
-  static const String tracks = 'Tracks';
-  static const String searchTracks = 'Search Tracks';
-  static const String playlistCreated = 'Playlist created successfully';
-  static const String trackAdded = 'Track added to playlist';
-  
-  static const String play = 'Play';
-  static const String pause = 'Pause';
-  static const String playPreview = 'Play Preview';
-  static const String pausePreview = 'Pause Preview';
-  static const String noPreviewAvailable = 'No preview available for this track';
-  
-  static const String friends = 'Friends';
-  static const String addFriend = 'Add Friend';
-  static const String deezer = 'Deezer';
-  static const String local = 'Local';
-  static const String addToLibrary = 'Add to Library';
-  static const String addedToLibrary = 'Added to your library';
-  static const String addToPlaylist = 'Add to Playlist';
-  
-  static const String loading = 'Loading...';
-  static const String error = 'Error';
-  static const String noResultsFound = 'No results found';
-  static const String noTracksFound = 'No tracks found';
-  static const String tryDifferentKeywords = 'Try searching with different keywords';
-  static const String connectionErrorMessage = 'Connection error. Check your internet.';
-  static const String featureComingSoon = 'This feature is coming soon!';
-  static const String searchForTracks = 'Search for tracks';
-  static const String deleteAccountWarning = 'This will permanently delete your account and all associated data.';
-  static const String confirmDelete = 'Are you sure you want to delete';
-}
-
-class Validators {
-  static String? required(String? value, String fieldName) {
-    if (value?.isEmpty ?? true) return 'Please enter $fieldName';
-    return null;
-  }
-
-  static String? email(String? value) {
-    if (value?.isEmpty ?? true) return 'Please enter an email';
-    if (!RegExp(r'^[\w-\.]+@([\w-]+\.)+[\w-]{2,4}$').hasMatch(value!)) {
-      return 'Please enter a valid email';
-    }
-    return null;
-  }
-
-  static String? password(String? value) {
-    if (value?.isEmpty ?? true) return 'Please enter password';
-    if (value!.length < AppConstants.minPasswordLength) {
-      return 'Password must be at least ${AppConstants.minPasswordLength} characters';
-    }
-    return null;
-  }
-
-  static String? playlistName(String? value) {
-    if (value?.isEmpty ?? true) return 'Please enter a playlist name';
-    return null;
-  }
-
-  static String? description(String? value) => null;
-
-  static String? username(String? value) {
-    if (value?.isEmpty ?? true) return 'Please enter a username';
-    if (value!.length < 3) return 'Username must be at least 3 characters';
-    return null;
-  }
-}
