+++ frontend-dart/lib/providers/dynamic_theme_provider.dart
@@ -2,7 +2,7 @@
 import 'package:flutter/material.dart';
 import 'package:palette_generator/palette_generator.dart';
 import 'package:cached_network_image/cached_network_image.dart';
-import '../core/app_core.dart';
+import '../core/consolidated_core.dart';
 
 class DynamicThemeProvider with ChangeNotifier {
   Color _primaryColor = AppTheme.primary;
