+++ frontend-dart/lib/screens/friends/friend_request_screen.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import '../../providers/auth_provider.dart';
 import '../../providers/friend_provider.dart';
-import '../../core/app_core.dart';
+import '../../core/consolidated_core.dart';
 import '../../widgets/common_widgets.dart';
 
 class FriendRequestScreen extends StatefulWidget {
