+++ frontend-dart/lib/screens/music/control_delegation_screen.dart
@@ -1,6 +1,6 @@
 // lib/screens/music/control_delegation_screen.dart
 import 'package:flutter/material.dart';
-import '../../core/app_core.dart';
+import '../../core/consolidated_core.dart';
 
 class MusicControlDelegationScreen extends StatelessWidget {
   const MusicControlDelegationScreen({Key? key}) : super(key: key);
