+++ frontend-dart/lib/screens/music/deezer_track_detail_screen.dart
@@ -1,6 +1,6 @@
 // lib/screens/music/deezer_track_detail_screen.dart
 import 'package:flutter/material.dart';
-import '../../core/app_core.dart';
+import '../../core/consolidated_core.dart';
 
 class DeezerTrackDetailScreen extends StatelessWidget {
   final String trackId;
