+++ frontend-dart/lib/screens/music/track_search_screen.dart
@@ -4,7 +4,7 @@ import 'package:provider/provider.dart';
 import '../../providers/music_provider.dart';
 import '../../providers/device_provider.dart';
 import '../../services/music_player_service.dart';
-import '../../core/app_core.dart';
+import '../../core/consolidated_core.dart';
 import '../../widgets/unified_components.dart';
 import '../../models/models.dart';
 import '../../utils/dialog_utils.dart';
