--- frontend-dart/lib/screens/profile/components/utils.dart
@@ -1,11 +1,0 @@
-// lib/screens/profile/components/utils.dart
-import 'package:intl/intl.dart';
-
-class Utils {
-
-  static String formatDate(DateTime? date) {
-    if (date != null) {
-      return DateFormat('yyyy-MM-dd').format(date);
-    }
-    return '';
-  }
