+++ frontend-dart/lib/screens/profile/user_password_change_screen.dart
@@ -1,7 +1,7 @@
 // lib/screens/profile/user_password_change_screen.dart
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import '../../core/app_core.dart';
+import '../../core/consolidated_core.dart';
 import '../../widgets/common_widgets.dart';
 import '../../providers/auth_provider.dart';
 import '../../providers/profile_provider.dart';
