+++ frontend-dart/lib/services/api_service.dart
@@ -4,7 +4,7 @@ import 'dart:io';
 import 'package:http/http.dart' as http;
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import '../models/models.dart';
-import '../core/app_core.dart';
+import '../core/consolidated_core.dart';
 
 class ApiService {
   static final ApiService _instance = ApiService._internal();
@@ -20,22 +20,33 @@ class ApiService {
 
   Future<T> _handleRequest<T>(
     Future<http.Response> Function() request,
-    T Function(Map<String, dynamic>) parser,
-  ) async {
+    T Function(dynamic) parser, {
+    bool expectJson = true,
+  }) async {
     try {
       final response = await request();
+      
       if (response.statusCode >= 400) {
         String errorMessage;
         try {
-          final errorData = json.decode(response.body);
-          errorMessage = errorData['error'] ?? errorData['detail'] ?? '${AppStrings.error}: ${response.statusCode}';
+          if (expectJson) {
+            final errorData = json.decode(response.body);
+            errorMessage = _extractErrorMessage(errorData, response.statusCode);
+          } else {
+            errorMessage = response.body.isNotEmpty ? response.body : 'HTTP ${response.statusCode}';
+          }
         } catch (e) {
           errorMessage = '${AppStrings.error}: ${response.statusCode}';
         }
         throw ApiException(errorMessage);
       }
-      final data = json.decode(response.body);
-      return parser(data);
+
+      if (expectJson && response.body.isNotEmpty) {
+        final data = json.decode(response.body);
+        return parser(data);
+      } else {
+        return parser(response.body);
+      }
     } on SocketException {
       throw ApiException(AppStrings.connectionErrorMessage);
     } catch (e) {
@@ -44,6 +55,18 @@ class ApiService {
     }
   }
 
+  String _extractErrorMessage(dynamic errorData, int statusCode) {
+    if (errorData is Map<String, dynamic>) {
+      return errorData['error'] ?? 
+             errorData['detail'] ?? 
+             errorData['message'] ?? 
+             '${AppStrings.error}: $statusCode';
+    } else if (errorData is String) {
+      return errorData;
+    }
+    return '${AppStrings.error}: $statusCode';
+  }
+  
   Future<AuthResult> login(String username, String password) async {
     return _handleRequest(
       () => http.post(
@@ -99,6 +122,173 @@ class ApiService {
     );
   }
 
+  Future<void> facebookLink(String? token, String fbAccessToken) async {
+    await _handleRequest(
+      () => http.post(
+        Uri.parse('$_baseUrl/auth/facebook/link/'),
+        headers: _getHeaders(token),
+        body: json.encode({'fbAccessToken': fbAccessToken}),
+      ),
+      (_) => null,
+    );
+  }
+
+  Future<void> googleLink(String type, String? token, String idToken) async {
+    await _handleRequest(
+      () => http.post(
+        Uri.parse('$_baseUrl/auth/google/link/'),
+        headers: _getHeaders(token),
+        body: json.encode({'idToken': idToken, 'type': type}),
+      ),
+      (_) => null,
+    );
+  }
+
+  Future<void> forgotPassword(String email) async {
+    await _handleRequest(
+      () => http.post(
+        Uri.parse('$_baseUrl/users/forgot_password/'),
+        headers: _getHeaders(),
+        body: json.encode({'email': email}),
+      ),
+      (_) => null,
+    );
+  }
+
+  Future<void> forgotChangePassword(String email, String otpStr, String password) async {
+    await _handleRequest(
+      () => http.post(
+        Uri.parse('$_baseUrl/users/forgot_change_password/'),
+        headers: _getHeaders(),
+        body: json.encode({'email': email, 'otp': int.parse(otpStr), 'password': password}),
+      ),
+      (_) => null,
+    );
+  }
+
+  Future<Map<String, dynamic>> getUser(String? token) async {
+    return _handleRequest(
+      () => http.get(Uri.parse('$_baseUrl/users/get_user/'), headers: _getHeaders(token)),
+      (data) => data,
+    );
+  }
+
+  Future<void> userPasswordChange(String? token, String currentPassword, String newPassword) async {
+    await _handleRequest(
+      () => http.post(
+        Uri.parse('$_baseUrl/users/user_password_change/'),
+        headers: _getHeaders(token),
+        body: json.encode({'currentPassword': currentPassword, 'newPassword': newPassword}),
+      ),
+      (_) => null,
+    );
+  }
+
+  Future<Map<String, dynamic>> getProfilePublic(String? token) async {
+    return _handleRequest(
+      () => http.get(Uri.parse('$_baseUrl/profile/public/'), headers: _getHeaders(token)),
+      (data) => data,
+    );
+  }
+
+  Future<Map<String, dynamic>> getProfilePrivate(String? token) async {
+    return _handleRequest(
+      () => http.get(Uri.parse('$_baseUrl/profile/private/'), headers: _getHeaders(token)),
+      (data) => data,
+    );
+  }
+
+  Future<Map<String, dynamic>> getProfileFriend(String? token) async {
+    return _handleRequest(
+      () => http.get(Uri.parse('$_baseUrl/profile/friend/'), headers: _getHeaders(token)),
+      (data) => data,
+    );
+  }
+
+  Future<Map<String, dynamic>> getProfileMusic(String? token) async {
+    return _handleRequest(
+      () => http.get(Uri.parse('$_baseUrl/profile/music/'), headers: _getHeaders(token)),
+      (data) => data,
+    );
+  }
+
+  Future<void> updateAvatar(String? token, String? avatarBase64, String? mimeType) async {
+    await _handleRequest(
+      () => http.post(
+        Uri.parse('$_baseUrl/profile/public/update/'),
+        headers: _getHeaders(token),
+        body: json.encode({'avatarBase64': avatarBase64, 'mimeType': mimeType}),
+      ),
+      (_) => null,
+    );
+  }
+
+  Future<void> updatePublicBasic(String? token, String? gender, String? location) async {
+    await _handleRequest(
+      () => http.post(
+        Uri.parse('$_baseUrl/profile/public/update/'),
+        headers: _getHeaders(token),
+        body: json.encode({'gender': gender, 'location': location}),
+      ),
+      (_) => null,
+    );
+  }
+
+  Future<void> updatePublicBio(String? token, String? bio) async {
+    await _handleRequest(
+      () => http.post(
+        Uri.parse('$_baseUrl/profile/public/update/'),
+        headers: _getHeaders(token),
+        body: json.encode({'bio': bio}),
+      ),
+      (_) => null,
+    );
+  }
+
+  Future<void> updatePrivateInfo(String? token, String? firstName, String? lastName, String? phone, String? street, String? country, String? postalCode) async {
+    await _handleRequest(
+      () => http.post(
+        Uri.parse('$_baseUrl/profile/private/update/'),
+        headers: _getHeaders(token),
+        body: json.encode({
+          'firstName': firstName,
+          'lastName': lastName,
+          'phone': phone,
+          'street': street,
+          'country': country,
+          'postalCode': postalCode,
+        }),
+      ),
+      (_) => null,
+    );
+  }
+
+  Future<void> updateFriendInfo(String? token, String? dob, List<String>? hobbies, String? friendInfo) async {
+    await _handleRequest(
+      () => http.post(
+        Uri.parse('$_baseUrl/profile/friend/update/'),
+        headers: _getHeaders(token),
+        body: json.encode({
+          'dob': dob,
+          'hobbies': hobbies,
+          'friendInfo': friendInfo,
+        }),
+      ),
+      (_) => null,
+    );
+  }
+
+  Future<void> updateMusicPreferences(String? token, List<String>? musicPreferences) async {
+    await _handleRequest(
+      () => http.post(
+        Uri.parse('$_baseUrl/profile/music/update/'),
+        headers: _getHeaders(token),
+        body: json.encode({'musicPreferences': musicPreferences}),
+      ),
+      (_) => null,
+    );
+  }
+
   Future<List<Playlist>> getPlaylists({required String token, bool publicOnly = false}) async {
     final endpoint = publicOnly ? '/playlists/public_playlists/' : '/playlists/saved_playlists/';
     return _handleRequest(
@@ -138,6 +328,62 @@ class ApiService {
     );
   }
 
+  Future<void> removeTrackFromPlaylist({
+    required String playlistId,
+    required String trackId,
+    required String token,
+    String? deviceUuid,
+  }) async {
+    await _handleRequest(
+      () => http.post(
+        Uri.parse('$_baseUrl/playlists/$playlistId/remove_tracks'),
+        headers: _getHeaders(token),
+        body: json.encode({
+          'track_id': trackId,
+          if (deviceUuid != null) 'device_uuid': deviceUuid,
+        }),
+      ),
+      (_) => null,
+    );
+  }
+
+  Future<void> moveTrackInPlaylist({
+    required String playlistId,
+    required int rangeStart,
+    required int insertBefore,
+    int rangeLength = 1,
+    required String token,
+  }) async {
+    await _handleRequest(
+      () => http.post(
+        Uri.parse('$_baseUrl/playlists/move-track/'),
+        headers: _getHeaders(token),
+        body: json.encode({
+          'playlist_id': int.parse(playlistId),
+          'range_start': rangeStart,
+          'insert_before': insertBefore,
+          'range_length': rangeLength,
+        }),
+      ),
+      (_) => null,
+    );
+  }
+
+  Future<void> inviteUserToPlaylist({
+    required String playlistId,
+    required int userId,
+    required String token,
+  }) async {
+    await _handleRequest(
+      () => http.post(
+        Uri.parse('$_baseUrl/playlists/$playlistId/invite-user/'),
+        headers: _getHeaders(token),
+        body: json.encode({'user_id': userId}),
+      ),
+      (_) => null,
+    );
+  }
+
   Future<List<Track>> searchTracks(String query, {bool deezer = true}) async {
     final endpoint = deezer ? '/deezer/search/' : '/tracks/search/';
     final param = deezer ? 'q' : 'query';
@@ -154,6 +400,16 @@ class ApiService {
     return _handleRequest(
       () => http.get(Uri.parse('$_baseUrl/deezer/track/$trackId/')),
       (data) => Track.fromJson(data),
+    );
+  }
+
+  Future<void> addTrackFromDeezer(String deezerTrackId, String token) async {
+    await _handleRequest(
+      () => http.post(
+        Uri.parse('$_baseUrl/tracks/add_from_deezer/$deezerTrackId/'),
+        headers: _getHeaders(token),
+      ),
+      (_) => null,
     );
   }
 
@@ -198,6 +454,30 @@ class ApiService {
     }
   }
 
+  Future<void> delegateDeviceControl({
+    required String deviceUuid,
+    required int delegateUserId,
+    required bool canControl,
+    required String token,
+  }) async {
+    await _handleRequest(
+      () => http.post(
+        Uri.parse('$_baseUrl/devices/delegate/'),
+        headers: _getHeaders(token),
+        body: json.encode({
+          'device_uuid': deviceUuid,
+          'delegate_user_id': delegateUserId,
+          'can_control': canControl,
+        }),
+      ),
+      (_) => null,
+    );
+  }
+
+  Future<List<Device>> getAllUserDevices(String token) async {
+    return getUserDevices(token); 
+  }
+
   Future<List<int>> getFriends(String token) async {
     return _handleRequest(
       () => http.get(Uri.parse('$_baseUrl/users/get_friends/'), headers: _getHeaders(token)),
@@ -216,153 +496,6 @@ class ApiService {
         body: json.encode({'user_id': userId}),
       ),
       (data) => data['message'] ?? 'Friend request sent',
-    );
-  }
-
-  Future<Map<String, dynamic>> getUser(String? token) async {
-    return _handleRequest(
-      () => http.get(Uri.parse('$_baseUrl/users/get_user/'), headers: _getHeaders(token)),
-      (data) => data,
-    );
-  }
-
-  Future<void> userPasswordChange(String? token, String currentPassword, String newPassword) async {
-    await _handleRequest(
-      () => http.post(
-        Uri.parse('$_baseUrl/users/user_password_change/'),
-        headers: _getHeaders(token),
-        body: json.encode({
-          'currentPassword': currentPassword,
-          'newPassword': newPassword,
-        }),
-      ),
-      (_) => null,
-    );
-  }
-
-  Future<void> facebookLink(String? token, String fbAccessToken) async {
-    await _handleRequest(
-      () => http.post(
-        Uri.parse('$_baseUrl/auth/facebook/link/'),
-        headers: _getHeaders(token),
-        body: json.encode({'fbAccessToken': fbAccessToken}),
-      ),
-      (_) => null,
-    );
-  }
-
-  Future<void> googleLink(String type, String? token, String idToken) async {
-    await _handleRequest(
-      () => http.post(
-        Uri.parse('$_baseUrl/auth/google/link/'),
-        headers: _getHeaders(token),
-        body: json.encode({'idToken': idToken, 'type': type}),
-      ),
-      (_) => null,
-    );
-  }
-
-  Future<Map<String, dynamic>> get(String endpoint, String token) async {
-    return _handleRequest(
-      () => http.get(Uri.parse('$_baseUrl$endpoint'), headers: _getHeaders(token)),
-      (data) => data,
-    );
-  }
-
-  Future<Map<String, dynamic>> post(String endpoint, Map<String, dynamic> body, String token) async {
-    return _handleRequest(
-      () => http.post(
-        Uri.parse('$_baseUrl$endpoint'),
-        headers: _getHeaders(token),
-        body: json.encode(body),
-      ),
-      (data) => data,
-    );
-  }
-
-  Future<void> moveTrackInPlaylist({
-    required String playlistId,
-    required int rangeStart,
-    required int insertBefore,
-    int rangeLength = 1,
-    required String token,
-  }) async {
-    return _handleRequest(
-      () => http.post(
-        Uri.parse('$_baseUrl/playlists/move-track/'),
-        headers: _getHeaders(token),
-        body: json.encode({
-          'playlist_id': int.parse(playlistId),
-          'range_start': rangeStart,
-          'insert_before': insertBefore,
-          'range_length': rangeLength,
-        }),
-      ),
-      (_) => null,
-    );
-  }
-
-  Future<void> inviteUserToPlaylist({
-    required String playlistId,
-    required int userId,
-    required String token,
-  }) async {
-    return _handleRequest(
-      () => http.post(
-        Uri.parse('$_baseUrl/playlists/$playlistId/invite-user/'),
-        headers: _getHeaders(token),
-        body: json.encode({'user_id': userId}),
-      ),
-      (_) => null,
-    );
-  }
-
-  Future<void> delegateDeviceControl({
-    required String deviceUuid,
-    required int delegateUserId,
-    required bool canControl,
-    required String token,
-  }) async {
-    return _handleRequest(
-      () => http.post(
-        Uri.parse('$_baseUrl/devices/delegate/'),
-        headers: _getHeaders(token),
-        body: json.encode({
-          'device_uuid': deviceUuid,
-          'delegate_user_id': delegateUserId,
-          'can_control': canControl,
-        }),
-      ),
-      (_) => null,
-    );
-  }
-
-  Future<List<Device>> getAllUserDevices(String token) async {
-    return _handleRequest(
-      () => http.get(Uri.parse('$_baseUrl/devices/'), headers: _getHeaders(token)),
-      (data) {
-        final devices = data['devices'] as List<dynamic>;
-        return devices.map((d) => Device.fromJson(d as Map<String, dynamic>)).toList();
-      },
-    );
-  }
-
-  Future<void> removeTrackFromPlaylist({
-    required String playlistId,
-    required String trackId,
-    required String token,
-    String? deviceUuid,
-  }) async {
-    return _handleRequest(
-      () => http.post(
-        Uri.parse('$_baseUrl/playlists/$playlistId/remove_tracks'),
-        headers: _getHeaders(token),
-        body: json.encode({
-          'track_id': trackId,
-          if (deviceUuid != null) 'device_uuid': deviceUuid,
-        }),
-      ),
-      (_) => null,
     );
   }
 
@@ -397,7 +530,7 @@ class ApiService {
   }
 
   Future<void> removeFriend(int friendId, String token) async {
-    return _handleRequest(
+    await _handleRequest(
       () => http.post(
         Uri.parse('$_baseUrl/users/remove_friend/$friendId/'),
         headers: _getHeaders(token),
@@ -406,14 +539,45 @@ class ApiService {
     );
   }
 
-  Future<void> addTrackFromDeezer(String deezerTrackId, String token) async {
+  Future<Map<String, dynamic>> get(String endpoint, String token) async {
+    return _handleRequest(
+      () => http.get(Uri.parse('$_baseUrl$endpoint'), headers: _getHeaders(token)),
+      (data) => data,
+    );
+  }
+
+  Future<Map<String, dynamic>> post(String endpoint, Map<String, dynamic> body, String token) async {
     return _handleRequest(
       () => http.post(
-        Uri.parse('$_baseUrl/tracks/add_from_deezer/$deezerTrackId/'),
+        Uri.parse('$_baseUrl$endpoint'),
         headers: _getHeaders(token),
+        body: json.encode(body),
       ),
-      (_) => null,
+      (data) => data,
     );
+  }
+
+  Future<bool> checkConnectivity() async {
+    try {
+      final response = await http.get(
+        Uri.parse('$_baseUrl/health'),
+        headers: _getHeaders(),
+      ).timeout(const Duration(seconds: 5));
+      return response.statusCode < 400;
+    } catch (e) {
+      return false;
+    }
+  }
+
+  Future<Map<String, dynamic>?> getApiInfo() async {
+    try {
+      return await _handleRequest(
+        () => http.get(Uri.parse('$_baseUrl/info')),
+        (data) => data,
+      );
+    } catch (e) {
+      return null;
+    }
   }
 }
 
@@ -429,7 +593,11 @@ class AuthResult {
 
 class ApiException implements Exception {
   final String message;
-  ApiException(this.message);
+  final int? statusCode;
+  final dynamic originalError;
+
+  ApiException(this.message, {this.statusCode, this.originalError});
+
   @override
   String toString() => message;
 }
